Deal with packed data.
A word on how I figured this out.

The .cat descrambling I found out by looking at Cycrows code.  I'm
pretty sure he (or whoever he learned this from) figured out the
algorithm by decompiling something (which was compiled with loop
unrolling) because it was jumping through all these weird hoops and
having all those state variables that were named like registers and 5
byte blocks and doing things backwards and all kinds of weirdness
that's completely unnecessary for something that is as simple as `data
^= ((219 + offset) % 256)`. You really don't need 20 lines of code
when one is more readable. Another, less charitable explanation is
that whoever wrote the scrambling (note how I very deliberately don't
call it encryption, because it isn't anything close to that) actually
thought that all these hoops actually do something, which they don't.

The handling of .pck data in Cycrows code does not match the one
cat/dat file I am testing this on (Mayhem mod). His xor dance has the
xor coookie as the first byte of data. My test data doesn't. Instead
it is consistently xor:ed with 51. Why 51? Beats me. I can't figure
out why anyone would even want to xor things in the first place
because it prevents mmap:ing (or windows equivalent) of the files,
which would be quite useful and would make things a bit less slow
(although with xml everywhere it is clear that performance was
definitely not a design goal).

The data is a gzip container. This is also figured out a bit
back-asswards from Cycrows code since he does all this song and dance
to skip bytes from some unclear header. I had a hunch that this was a
gzip header, so going on the assumption that it is, I xor:ed the first
two bytes of the data with 31 and 139 and both times it gave
51. Xor:ing the rest of the data with 51 allowed me to hand it over to
gzip to deal with the headers.

Is it likely that this will break down on other containers? Only
testing will tell. For now, I'm just trying to decode Mayhem, we'll
deal with the rest later.