I mean, really?
We can gain a 10% performance increase of treating the compiler
like it had certain developmental challenges.

I mean...

1. In a loop that does `for i := 0; i < x; i++ { slice[i] ... }`,
wouldn't it be reasonable to expect the compiler to do the bounds
check just by comparing x to len(slice) once instead of checking
`i` every time? I could change this loop to `for i := range slice`
in one of the two cases because it doesn't hurt to write to the
slice on short reads. The other loop can't be changed that trivally.

2. In a loop that does `for i := range x { x[i] ^= struct.member }`
Where exactly is the code that can change `struct.member` between
the iterations of the loop? Do we really have to read it from the
struct on each iteration? I haven't done these kinds of manual
optimizations for over two decades.